resume-extract/requirements.txt
fastapi==0.116.1
httpx==0.28.1
pdf2image==1.17.0
pypdf==5.4.0
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.1.1
uvicorn==0.35.0
